requirements.txt
certifi==2018.1.18
chardet==3.0.4
click==6.7
dateparser==0.7.0
Flask==0.12.2
Flask-Cors==3.0.3
gunicorn==19.7.1
html2text==2018.1.9
humanize==0.5.1
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
maya==0.3.3
pendulum==1.4.2
python-dateutil==2.6.1
python-dotenv==0.7.1
pytz==2018.3
pytzdata==2018.3
regex==2018.2.21
requests==2.18.4
ruamel.yaml==0.15.35
six==1.11.0
tzlocal==1.5.1
urllib3==1.22
Werkzeug==0.14.1
